completed merge of upstream fixes PhilippeDev/philippe_fixes in frippon/freeture

+++ src/Device.h
@@ -117,7 +117,7 @@ class Device {
         int         mNightGain;
         int         mDayExposure;
         int         mDayGain;
-        int         mFPS;
+        double      mFPS;
         int         mCamID;         // ID in a specific sdk.
         int         mGenCamID;      // General ID.
         Camera      *mCam;
